requirements.txt
feedparser>=6.0.10
requests>=2.31.0
pyyaml>=6.0.1
jinja2>=3.1.2
tqdm>=4.66.0
anthropic>=0.40.0
pypdf>=4.0.0
python-dotenv>=1.0.0
